% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTests.R
\name{calcDiff}
\alias{calcDiff}
\title{Calculate the difference in deuteration between two states for each peptide at each labelling time point}
\usage{
calcDiff(dataset, groups = NULL)
}
\arguments{
\item{dataset}{the HDXMS dataset in a long format with following columns: peptide, label_time, state and deuterium_level}

\item{groups}{indicator variable that defines the two states to be compared}
}
\value{
A data frame with the calculated differences for each peptide at each time point.
}
\description{
Calculate the difference in deuteration between two states for each peptide at each labelling time point
}
\examples{

#Please check the vignette of the HDXstats package.

}
