% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{calculateDeuteriumLevels}
\alias{calculateDeuteriumLevels}
\title{Convert output from vendor software to a format suitable for analysis with
HDXstats}
\usage{
calculateDeuteriumLevels(dataset, type = "DynamX cluster")
}
\arguments{
\item{dataset}{The HDXMS dataset from an external software tool. Please
make sure that for each peptide the same number of replicates are
reported. If not the global threshold, hybrid significance test and the
Ttest produce unreliable results}

\item{type}{Which software tool has been used to extract information from
the spectra. The default value is DynamX cluster.}
}
\value{
A data frame with columns: peptide, start, end, label_time, state and
deuterium_level
}
\description{
Convert output from vendor software to a format suitable for analysis with
HDXstats
}
\examples{
#Please check the vignette of the HDXstats package.
}
