% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convertLongToWide}
\alias{convertLongToWide}
\title{Convert the processed HDXMS dataset in long format to the wide format which is required for the moderated test-statistics.}
\usage{
convertLongToWide(datasetLong, nbRep)
}
\arguments{
\item{datasetLong}{The HDXMS dataset in long format with following columns: peptide, label_time, state and deuterium_level}

\item{nbRep}{The number of replicates. Make sure that each peptide has exactly the same number of replicates as specified here.}
}
\value{
The HDXMS dataset in wide format
}
\description{
Convert the processed HDXMS dataset in long format to the wide format which is required for the moderated test-statistics.
}
\examples{
#Please check the vignette of the HDXstats package.
}
