% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{coveragePlot}
\alias{coveragePlot}
\title{create a protein coverage plot}
\usage{
coveragePlot(dataset)
}
\arguments{
\item{dataset}{TThe HDXMS dataset from an external software tool with at least following columns: Sequence, Start, and End}
}
\value{
A ggplot object
}
\description{
create a protein coverage plot
}
\examples{
#Please check the vignette of the HDXstats package.
}
