% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createBarPlot}
\alias{createBarPlot}
\title{create a barplot showing the number of differential deuterated peptides}
\usage{
createBarPlot(
  diffDeutResult,
  alpha = 0.05,
  col = NULL,
  xlabels = NULL,
  groupedByTime = TRUE
)
}
\arguments{
\item{diffDeutResult}{output from a moderated test-statistics}

\item{alpha}{significance-level. Default is 0.05}

\item{col}{a vector of colors for the bars. Default is NULL}

\item{xlabels}{a vector of names for the labels of the x-axis. Default is NULL}

\item{groupedByTime}{logical variable indicating if the bars should be organized by labeling time. Default is TRUE}
}
\value{
A ggplot object
}
\description{
create a barplot showing the number of differential deuterated peptides
}
\examples{
#Please check the vignette of the HDXstats package.
}
