% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createButterflyPlot}
\alias{createButterflyPlot}
\title{create a butterfly-plot}
\usage{
createButterflyPlot(dataset, states = NULL)
}
\arguments{
\item{dataset}{output from calculateDeuteriumLevels}

\item{states}{vector indicating which protein states should be compared}
}
\value{
A ggplot object
}
\description{
create a butterfly-plot
}
\examples{
#Please check the vignette of the HDXstats package.
}
