% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{createContrastMatrix}
\alias{createContrastMatrix}
\title{Create the contrast matrix for the moderated test statistics}
\usage{
createContrastMatrix(states, label_time, designMatrix)
}
\arguments{
\item{states}{character vector specifying all protein states of interest}

\item{label_time}{character vector specifying all time points of interest}

\item{designMatrix}{matrix that corresponds to the design of the HDX-MS experiment}
}
\value{
Matrix which columns correspond to contrasts
}
\description{
Create the contrast matrix for the moderated test statistics
}
\examples{
#Please check the vignette of the HDXstats package.
}
