% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createDeuterationPlot}
\alias{createDeuterationPlot}
\title{create a deuteration-plot}
\usage{
createDeuterationPlot(dataset, peptides = NULL)
}
\arguments{
\item{dataset}{output from calculateDeuteriumLevels}

\item{peptides}{vector indicating which peptides should be shown}
}
\value{
A ggplot object
}
\description{
create a deuteration-plot
}
\examples{
#Please check the vignette of the HDXstats package.
}
