% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createMAPlot}
\alias{createMAPlot}
\title{create a MA-plot showing the difference in deuteration versus the average deuteration-value}
\usage{
createMAPlot(diffDeutResult, alpha = 0.05)
}
\arguments{
\item{diffDeutResult}{data.frame from one of the moderated test-statistics}

\item{alpha}{significance-level. Default is 0.05}
}
\value{
A ggplot object
}
\description{
create a MA-plot showing the difference in deuteration versus the average deuteration-value
}
\examples{
#Please check the vignette of the HDXstats package.
}
