% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createVolcanoPlot}
\alias{createVolcanoPlot}
\title{create a volcano plot from the output of the moderated test-statistics}
\usage{
createVolcanoPlot(diffDeutResult, alpha = 0.05)
}
\arguments{
\item{diffDeutResult}{data.frame from one of the moderated test-statistics}

\item{alpha}{significance-level. Default is 0.05}
}
\value{
A ggplot object
}
\description{
create a volcano plot from the output of the moderated test-statistics
}
\examples{
#Please check the vignette of the HDXstats package.
}
