% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{createWoodsPlot}
\alias{createWoodsPlot}
\title{create a Woods plot}
\usage{
createWoodsPlot(diffDeutResult, seqInfo, alpha = 0.05)
}
\arguments{
\item{diffDeutResult}{data.frame from one of the moderated test-statistics}

\item{seqInfo}{data.frame containing peptide sequence, start position and end position}

\item{alpha}{significance level}
}
\value{
A ggplot object
}
\description{
create a Woods plot
}
\examples{
#Please check the vignette of the HDXstats package.
}
