% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{exploratoryDataAnalysis}
\alias{exploratoryDataAnalysis}
\title{Create a principal component plot or spectral map for HDX-MS deuteration levels}
\usage{
exploratoryDataAnalysis(dataset, nbRep = NULL, type = "sma")
}
\arguments{
\item{dataset}{The HDXMS dataset in long format with at least following columns: peptide, label_time, state and deuterium_level}

\item{nbRep}{The number of replicates. Make sure that each peptide has exactly the same number of replicates as specified here.}

\item{type}{The type of plot: principal component (pca) or spectral map (sma). Default value is sma.}
}
\value{
A plot.mpm object
}
\description{
Create a principal component plot or spectral map for HDX-MS deuteration levels
}
\examples{
#Please check the vignette of the HDXstats package.
}
