% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTests.R
\name{fixedThreshold}
\alias{fixedThreshold}
\title{Compares the difference in deuterium-levels between two states against a fixed user-defined threshold}
\usage{
fixedThreshold(dataset, nbRep, group, threshold)
}
\arguments{
\item{dataset}{The HDXMS dataset in a long format with following columns: peptide, label_time, state and deuterium_level.}

\item{nbRep}{The number of replicates. Make sure that each peptide has exactly the same number of replicates as specified here.}

\item{group}{Which protein states need to be compared with each other.}

\item{threshold}{The minimum difference in deuterium-levels that is considered to be significant.}
}
\value{
A list with the results of the fixed threshold approach. Each element of the list corresponds to a specific comparison
}
\description{
Compares the difference in deuterium-levels between two states against a fixed user-defined threshold
}
\examples{

#Please check the vignette of the HDXstats package.

}
