% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTests.R
\name{ttestHDX}
\alias{ttestHDX}
\title{Calculates the two sample t-test for differential HDX-MS data}
\usage{
ttestHDX(dataset, nbRep, group, alpha = 0.05)
}
\arguments{
\item{dataset}{The HDXMS dataset in a long format with following columns: peptide, label_time, state and deuterium_level}

\item{nbRep}{The number of replicates. Make sure that each peptide has exactly the same number of replicates as specified here.}

\item{group}{Which protein states need to be compared with each other.}

\item{alpha}{Significance level. Default is 0.05}
}
\value{
A list with the results of the two sample t-test. Each element of the list corresponds to a specific comparison
}
\description{
Calculates the two sample t-test for differential HDX-MS data
}
\examples{

#Please check the vignette of the HDXstats package.

}
